Added 2 files - conftest.py, data.py Completed 20 tests
