+++ data.py
@@ -1,0 +1,6 @@
+''' Константы для тестов tests.py '''
+
+BOOKS_COLLECTION = ['Книга 1', 'Книга 2', 'Книга 3', 'Книга 4', 'Книга 5']
+GENRE_LIST = ['Фантастика', 'Ужасы', 'Детективы', 'Мультфильмы', 'Комедии']
+GENRE_AGE_RATING = ['Ужасы', 'Детективы']
+GENRE_FOR_CHLDREN = ['Фантастика', 'Мультфильмы', 'Комедии']
